v0.1.2/T07: refactor combined_server.py into server/combined.py with shim
